frdm-k64f: Move clock initialization to board_init
The clock setting on this board depends on the existence of an external
RMII crystal, which is not mandatory for the CPU itself.
